wip: tidy
