+++ webui/src/ImportExport/Import/Full.jsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export function ImportFullWizard({ snapshot, instanceRemap }) {
+	return <p>Full2</p>
+}
